move overlay code to debug-tools file, add doc comment and add function to help
